# Conflicts:
#	src/main/java/com/mackenzie/br/socialmedia/controller/RecommendationController.java
#	src/main/java/com/mackenzie/br/socialmedia/controller/SearchController.java
#	src/main/java/com/mackenzie/br/socialmedia/service/PublicationService.java

+++ src/main/java/com/mackenzie/br/socialmedia/domain/Professional_FriendsDomain.java
@@ -4,7 +4,7 @@ public class Professional_FriendsDomain {
 	
 	private ProfessionalDomain professional;
 	
-	private Long numberOfFriends;
+	private Integer numberOfFriends;
 
 	public ProfessionalDomain getProfessional() {
 		return professional;
@@ -14,11 +14,11 @@ public class Professional_FriendsDomain {
 		this.professional = professional;
 	}
 
-	public Long getNumberOfFriends() {
+	public Integer getNumberOfFriends() {
 		return numberOfFriends;
 	}
 
-	public void setNumberOfFriends(Long numberOfFriends) {
+	public void setNumberOfFriends(Integer numberOfFriends) {
 		this.numberOfFriends = numberOfFriends;
 	} 
 }
